refactor: rename prisma module && remove useless imports
