+++ src/album/album.module.ts
@@ -1,10 +1,8 @@
 import { Module } from '@nestjs/common';
 import { AlbumService } from './album.service';
 import { AlbumController } from './album.controller';
-import { RepositoryModule } from 'src/repository/repository.module';
 
 @Module({
-  imports: [RepositoryModule],
   controllers: [AlbumController],
   providers: [AlbumService],
 })
