+++ src/artist/artist.module.ts
@@ -1,10 +1,8 @@
 import { Module } from '@nestjs/common';
 import { ArtistService } from './artist.service';
 import { ArtistController } from './artist.controller';
-import { RepositoryModule } from 'src/repository/repository.module';
 
 @Module({
-  imports: [RepositoryModule],
   controllers: [ArtistController],
   providers: [ArtistService],
 })
