+++ src/track/track.module.ts
@@ -1,10 +1,8 @@
 import { Module } from '@nestjs/common';
 import { TrackService } from './track.service';
 import { TrackController } from './track.controller';
-import { RepositoryModule } from 'src/repository/repository.module';
 
 @Module({
-  imports: [RepositoryModule],
   controllers: [TrackController],
   providers: [TrackService],
 })
