+++ src/user/user.module.ts
@@ -1,10 +1,8 @@
 import { Module } from '@nestjs/common';
 import { UserService } from './user.service';
 import { UserController } from './user.controller';
-import { RepositoryModule } from 'src/repository/repository.module';
 
 @Module({
-  imports: [RepositoryModule],
   controllers: [UserController],
   providers: [UserService],
 })
